requirements.txt
wxPython>=4.2.0
python-mpv>=1.0.0
jellyfin-apiclient-python>=1.9.0
